requirements.txt
pystray>=0.19.3
psutil>=5.9.0
Pillow>=9.0.0
requests>=2.28.0
